requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.2.5
django-cors-headers==4.7.0
djangorestframework==3.15.2
gunicorn==23.0.0
idna==3.10
packaging==25.0
pillow==11.1.0
python-dotenv==1.1.0
python-slugify==8.0.4
requests==2.32.3
sqlparse==0.5.3
stripe==12.4.0
text-unidecode==1.3
typing_extensions==4.14.1
urllib3==2.5.0
